+++ internal/tests/features/registry_register_named_test.go
@@ -56,6 +56,33 @@ func Test_Registry_register_named_service_consume_factory(t *testing.T) {
 	assert.NotNil(t, actual.localDataService)
 }
 
+func Test_Registry_register_named_service_resolve_as_list(t *testing.T) {
+
+	// Arrange
+	registry := registration.NewServiceRegistry()
+	_ = features.RegisterNamed[dataService](registry,
+		registration.NamedServiceRegistration("remote", newRemoteDataService, types.LifetimeSingleton),
+		registration.NamedServiceRegistration("local", newLocalDataService, types.LifetimeTransient))
+
+	resolver := resolving.NewResolver(registry)
+	scopedContext := resolving.NewScopedContext(context.Background())
+
+	// Act
+	actual, err := resolving.ResolveRequiredServices[dataService](resolver, scopedContext)
+	namedServiceFactory, _ := resolving.ResolveRequiredService[func(string) (dataService, error)](resolver, scopedContext)
+	remote, _ := namedServiceFactory("remote")
+	local, _ := namedServiceFactory("local")
+
+	// Assert
+	assert.NoError(t, err)
+	assert.NotNil(t, actual)
+	assert.Equal(t, 2, len(actual))
+	assert.NotNil(t, remote)
+	assert.Equal(t, "data from remote service", remote.FetchData())
+	assert.NotNil(t, local)
+	assert.Equal(t, "data from local service", local.FetchData())
+}
+
 type dataService interface {
 	FetchData() string
 }